requirements.txt
tk
Pillow
pandas
pyarrow